Rework the module structure.
This changes the location of all `RustlerFriendly` modules in the module
tree. Instead of being located in

`Membrane.VideoCompositor.Scene.RustlerFriendly.X`,

they're now located in modules

`Membrane.VideoCompositor.Scene.Object.X.RustlerFriendly`

Additionally, these modules are now defined in the same file their
parent module is in, which allows keeping all relevant parts in the same
file.
