+++ lib/membrane/video_compositor/scene/object.ex
@@ -5,6 +5,20 @@ defmodule Membrane.VideoCompositor.Scene.Object do
 
   alias Membrane.VideoCompositor.Scene.Object.{InputVideo, Layout, Texture}
 
+  defmodule RustlerFriendly do
+    @moduledoc false
+
+    alias Membrane.VideoCompositor.Scene.Object.InputVideo.RustlerFriendly, as: RFInputVideo
+    alias Membrane.VideoCompositor.Scene.Object.Layout.RustlerFriendly, as: RFLayout
+    alias Membrane.VideoCompositor.Scene.Object.Texture.RustlerFriendly, as: RFTexture
+
+    @type name :: Membrane.VideoCompositor.Scene.Object.name()
+
+    @type t :: {:layout, RFLayout.t()} | {:texture, RFTexture.t()} | {:video, RFInputVideo.t()}
+
+    @type object_output_resolution :: Texture.output_resolution() | Layout.output_resolution()
+  end
+
   @typedoc """
   Objects are renderable entities in VC, that can serve as input for other
   objects or as an output of the video.
@@ -37,9 +51,11 @@ defmodule Membrane.VideoCompositor.Scene.Object do
   (e.g. for corners rounding - same as input,
   for cropping - accordingly smaller than input)
   """
+  # FIXME: This and Texture.output_resolution should be reworked in some way.
+  #        I don't know what these should look like on the rust side.
   @type object_output_resolution :: Texture.output_resolution() | Layout.output_resolution()
 
-  @spec encode(t()) :: Membrane.VideoCompositor.Scene.RustlerFriendly.Object.t()
+  @spec encode(t()) :: RustlerFriendly.t()
   def encode(object) do
     case object do
       %InputVideo{} -> {:video, InputVideo.encode(object)}
