--- lib/membrane/video_compositor/scene/rustler_friendly/input_video.ex
@@ -1,10 +1,0 @@
-defmodule Membrane.VideoCompositor.Scene.RustlerFriendly.InputVideo do
-  @moduledoc false
-
-  @type t :: %__MODULE__{
-          input_pad: Membrane.Pad.ref_t()
-        }
-
-  @enforce_keys [:input_pad]
-  defstruct @enforce_keys
-end
